requirements.txt
Flask~=1.1.2
Flask-API.yandex~=0.6.2.1